+++ lib/src/components/card/impaktfull_card.dart
@@ -26,26 +26,30 @@ class ImpaktfullCard extends StatelessWidget {
             isSelected ? theme.borders.selectedCard : theme.borders.card;
         final shadow =
             isSelected ? theme.shadows.selectedCard : theme.shadows.card;
-        return Container(
-          decoration: BoxDecoration(
-            color: theme.colors.card,
-            border: border,
-            borderRadius:
-                BorderRadius.circular(theme.dimens.generalBorderRadius),
-            boxShadow: [
-              if (shadow != null) shadow,
-            ],
-          ),
-          margin: margin,
-          padding: EdgeInsets.only(
-            top: 16 - (border?.top.width ?? 0),
-            right: 16 - (border?.right.width ?? 0),
-            bottom: 16 - (border?.bottom.width ?? 0),
-            left: 16 - (border?.left.width ?? 0),
-          ).add(padding ?? EdgeInsets.zero),
-          child: ImpaktfullTouchFeedback(
-            onTap: onTap,
-            child: child,
+        return Padding(
+          padding: margin ?? EdgeInsets.zero,
+          child: Container(
+            decoration: BoxDecoration(
+              color: theme.colors.card,
+              border: border,
+              borderRadius:
+                  BorderRadius.circular(theme.dimens.generalBorderRadius),
+              boxShadow: [
+                if (shadow != null) shadow,
+              ],
+            ),
+            child: ImpaktfullTouchFeedback(
+              onTap: onTap,
+              child: Padding(
+                padding: EdgeInsets.only(
+                  top: 16 - (border?.top.width ?? 0),
+                  right: 16 - (border?.right.width ?? 0),
+                  bottom: 16 - (border?.bottom.width ?? 0),
+                  left: 16 - (border?.left.width ?? 0),
+                ).add(padding ?? EdgeInsets.zero),
+                child: child,
+              ),
+            ),
           ),
         );
       },