+++ lib/src/components/list_item/impaktfull_switch_list_item.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:impaktfull_ui/src/components/icon/impaktfull_svg_icon.dart';
-import 'package:impaktfull_ui/src/components/list_item/base_list_item.dart';
+import 'package:impaktfull_ui/src/components/list_item/impaktfull_simple_list_item.dart';
 import 'package:impaktfull_ui/src/components/switch/impaktfull_switch.dart';
 import 'package:impaktfull_ui/src/theme/impaktfull_theme_localizer.dart';
 
@@ -9,12 +9,12 @@ class ImpaktfullSwitchListItem extends StatelessWidget {
   final String? subTitle;
   final String? leadingAsset;
   final bool value;
-  final ValueChanged<bool> onChanged;
+  final ValueChanged<bool>? onChanged;
 
   const ImpaktfullSwitchListItem({
     required this.title,
     required this.value,
-    required this.onChanged,
+    this.onChanged,
     this.subTitle,
     this.leadingAsset,
     super.key,
@@ -23,20 +23,23 @@ class ImpaktfullSwitchListItem extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ImpaktfullThemeLocalizer(
-      builder: (context, theme) => BaseListItem(
-        title: title,
-        subTitle: subTitle,
-        onTap: () => onChanged(!value),
-        leadingWidget: leadingAsset == null
-            ? null
-            : ImpaktfullSvgIcon(
-                asset: leadingAsset!,
-                size: 16,
-                color: theme.colors.primary,
-              ),
-        trailingWidget: ImpaktfullSwitch(
-          value: value,
-          onChanged: onChanged,
+      builder: (context, theme) => IgnorePointer(
+        ignoring: onChanged == null,
+        child: ImpaktfullSimpleListItem(
+          title: title,
+          subTitle: subTitle,
+          onTap: () => onChanged?.call(!value),
+          leadingWidget: leadingAsset == null
+              ? null
+              : ImpaktfullSvgIcon(
+                  asset: leadingAsset!,
+                  size: 16,
+                  color: theme.colors.primary,
+                ),
+          trailingWidget: ImpaktfullSwitch(
+            value: value,
+            onChanged: onChanged,
+          ),
         ),
       ),
     );