+++ lib/src/components/separated_column/impaktfull_separated_column.dart
@@ -32,7 +32,7 @@ class ImpaktfullSeparatedColumn extends StatelessWidget {
                   ? Colors.transparent
                   : theme.colors.card,
             ),
-            child: Column(
+            child: ImpaktfullAutoLayout.vertical(
               children: children.isEmpty
                   ? []
                   : [