+++ lib/src/components/switch/impaktfull_switch.dart
@@ -2,7 +2,7 @@ import 'package:impaktfull_ui/impaktfull_ui.dart';
 
 class ImpaktfullSwitch extends StatelessWidget {
   final bool value;
-  final ValueChanged<bool> onChanged;
+  final ValueChanged<bool>? onChanged;
 
   const ImpaktfullSwitch({
     required this.value,
@@ -14,43 +14,47 @@ class ImpaktfullSwitch extends StatelessWidget {
   Widget build(BuildContext context) {
     return ImpaktfullThemeLocalizer(
       builder: (context, theme) => Center(
-        child: ImpaktfullTouchFeedback(
-          onTap: () => onChanged(!value),
-          color: theme.colors.card,
-          borderRadius: BorderRadius.circular(theme.dimens.switchBorderRadius),
-          child: AnimatedContainer(
-            duration: theme.durations.short,
-            curve: Curves.easeInOut,
-            decoration: BoxDecoration(
-              borderRadius:
-                  BorderRadius.circular(theme.dimens.switchBorderRadius),
-              border: Border.all(
-                color: value
-                    ? theme.colors.accent1
-                    : theme.colors.accent1TurnedOffState,
-                width: theme.dimens.borderWidth,
-              ),
-            ),
-            child: AnimatedPadding(
-              padding: EdgeInsets.only(
-                top: 4,
-                bottom: 4,
-                right: value ? 4 : 20,
-                left: value ? 20 : 4,
-              ),
-              curve: Curves.easeInOut,
+        child: Opacity(
+          opacity: onChanged == null ? 0.5 : 1,
+          child: ImpaktfullTouchFeedback(
+            onTap: onChanged == null ? null : () => onChanged!.call(!value),
+            color: theme.colors.card,
+            borderRadius:
+                BorderRadius.circular(theme.dimens.switchBorderRadius),
+            child: AnimatedContainer(
               duration: theme.durations.short,
-              child: AnimatedContainer(
-                duration: theme.durations.short,
-                curve: Curves.easeInOut,
-                height: 16,
-                width: 16,
-                decoration: BoxDecoration(
+              curve: Curves.easeInOut,
+              decoration: BoxDecoration(
+                borderRadius:
+                    BorderRadius.circular(theme.dimens.switchBorderRadius),
+                border: Border.all(
                   color: value
                       ? theme.colors.accent1
                       : theme.colors.accent1TurnedOffState,
-                  borderRadius: BorderRadius.circular(
-                    theme.dimens.switchThumbBorderRadius,
+                  width: theme.dimens.borderWidth,
+                ),
+              ),
+              child: AnimatedPadding(
+                padding: EdgeInsets.only(
+                  top: 4,
+                  bottom: 4,
+                  right: value ? 4 : 20,
+                  left: value ? 20 : 4,
+                ),
+                curve: Curves.easeInOut,
+                duration: theme.durations.short,
+                child: AnimatedContainer(
+                  duration: theme.durations.short,
+                  curve: Curves.easeInOut,
+                  height: 16,
+                  width: 16,
+                  decoration: BoxDecoration(
+                    color: value
+                        ? theme.colors.accent1
+                        : theme.colors.accent1TurnedOffState,
+                    borderRadius: BorderRadius.circular(
+                      theme.dimens.switchThumbBorderRadius,
+                    ),
                   ),
                 ),
               ),